Clock, DigitalClock, Meters and Gauges refactored.
