+++ org.csstudio.display.builder.representation.javafx/src/org/csstudio/display/builder/representation/javafx/widgets/MeterRepresentation.java
@@ -9,26 +9,15 @@
 package org.csstudio.display.builder.representation.javafx.widgets;
 
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.csstudio.display.builder.model.DirtyFlag;
 import org.csstudio.display.builder.model.WidgetProperty;
-import org.csstudio.display.builder.model.persist.NamedWidgetColors;
-import org.csstudio.display.builder.model.persist.WidgetColorService;
-import org.csstudio.display.builder.model.util.VTypeUtil;
 import org.csstudio.display.builder.model.widgets.MeterWidget;
 import org.csstudio.display.builder.model.widgets.MeterWidget.KnobPosition;
 import org.csstudio.display.builder.model.widgets.MeterWidget.Skin;
-import org.csstudio.display.builder.representation.javafx.JFXUtil;
-import org.diirt.vtype.Display;
-import org.diirt.vtype.VType;
-import org.diirt.vtype.ValueUtil;
 
 import eu.hansolo.medusa.Gauge;
-import eu.hansolo.medusa.GaugeBuilder;
 import eu.hansolo.medusa.Section;
 import eu.hansolo.medusa.TickLabelLocation;
 import javafx.geometry.Orientation;
@@ -40,64 +29,20 @@ import javafx.scene.paint.Color;
  * @author Claudio Rosati, European Spallation Source ERIC
  * @version 1.0.0 25 Jan 2017
  */
-public class MeterRepresentation extends RegionBaseRepresentation<Gauge, MeterWidget> {
+public class MeterRepresentation extends BaseGaugeRepresentation<MeterWidget> {
 
-    private static final Color MINOR_COLOR       = JFXUtil.convert(WidgetColorService.getColor(NamedWidgetColors.ALARM_MINOR));
-    private static final Color MINOR_COLOR_LIGHT = MINOR_COLOR.deriveColor(0.0, 1.0, 1.0, 0.2);
-    private static final Color MAJOR_COLOR       = JFXUtil.convert(WidgetColorService.getColor(NamedWidgetColors.ALARM_MAJOR));
-    private static final Color MAJOR_COLOR_LIGHT = MAJOR_COLOR.deriveColor(0.0, 1.0, 1.0, 0.2);
-
-    private final DirtyFlag          dirtyBehavior  = new DirtyFlag();
-    private final DirtyFlag          dirtyGeometry  = new DirtyFlag();
     private final DirtyFlag          dirtyLimits    = new DirtyFlag();
     private final DirtyFlag          dirtyLook      = new DirtyFlag();
-    private final DirtyFlag          dirtyValue     = new DirtyFlag();
-    private volatile double          high           = Double.NaN;
-    private volatile double          hihi           = Double.NaN;
-    private volatile double          lolo           = Double.NaN;
-    private volatile double          low            = Double.NaN;
-    private volatile double          max            = 100.0;
-    private volatile double          min            = 0.0;
     private MeterWidget.Skin         skin           = null;
     private MeterWidget.KnobPosition knobPosition   = null;
-    private final AtomicBoolean      updatingValue  = new AtomicBoolean(false);
     private volatile boolean         zonesHighlight = true;
 
     @Override
     public void updateChanges ( ) {
 
+        super.updateChanges();
+
         Object value;
-
-        if ( dirtyBehavior.checkAndClear() ) {
-
-            value = model_widget.propAnimated().getValue();
-
-            if ( !Objects.equals(value, jfx_node.isAnimated()) ) {
-                jfx_node.setAnimated((boolean) value);
-            }
-
-            value = model_widget.propAnimationDuration().getValue();
-
-            if ( !Objects.equals(value, jfx_node.getAnimationDuration()) ) {
-                jfx_node.setAnimationDuration((long) value);
-            }
-
-        }
-
-        if ( dirtyGeometry.checkAndClear() ) {
-
-            value = model_widget.propVisible().getValue();
-
-            if ( !Objects.equals(value, jfx_node.isVisible()) ) {
-                jfx_node.setVisible((boolean) value);
-            }
-
-            jfx_node.setLayoutX(model_widget.propX().getValue());
-            jfx_node.setLayoutY(model_widget.propY().getValue());
-            jfx_node.setPrefWidth(model_widget.propWidth().getValue());
-            jfx_node.setPrefHeight(model_widget.propHeight().getValue());
-
-        }
 
         if ( dirtyLook.checkAndClear() ) {
 
@@ -122,14 +67,10 @@ public class MeterRepresentation extends RegionBaseRepresentation<Gauge, MeterWi
                         break;
                 }
 
-                jfx_node.setSkinType(skinType);
-                jfx_node.setPrefWidth(model_widget.propWidth().getValue());
-                jfx_node.setPrefHeight(model_widget.propHeight().getValue());
-                jfx_node.setAutoScale(true);
-                jfx_node.setCheckAreasForValue(false);
-                jfx_node.setCheckSectionsForValue(false);
-                jfx_node.setCheckThreshold(false);
-                jfx_node.setLedVisible(false);
+                changeSkin(skinType);
+
+                jfx_node.setHighlightSections(zonesHighlight);
+                jfx_node.setKnobPosition(Pos.valueOf(knobPosition.name()));
                 jfx_node.setTickLabelLocation(TickLabelLocation.INSIDE);
 
                 switch ( skin ) {
@@ -159,72 +100,21 @@ public class MeterRepresentation extends RegionBaseRepresentation<Gauge, MeterWi
 
             }
 
-            value = model_widget.propTitle().getValue();
-
-            if ( !Objects.equals(value, jfx_node.getTitle()) ) {
-                jfx_node.setTitle((String) value);
-            }
-
-            value = JFXUtil.convert(model_widget.propTitleColor().getValue());
-
-            if ( !Objects.equals(value, jfx_node.getTitleColor()) ) {
-                jfx_node.setTitleColor((Color) value);
-            }
-
         }
 
         if ( dirtyLimits.checkAndClear() ) {
             jfx_node.setHighlightSections(zonesHighlight);
-            jfx_node.setMaxValue(max);
-            jfx_node.setMinValue(min);
-            jfx_node.setSectionsVisible(areZonesVisible());
-            jfx_node.setSections(createZones());
         }
 
-        if ( dirtyValue.checkAndClear() && updatingValue.compareAndSet(false, true) ) {
-            try {
-
-                final VType vtype = model_widget.runtimePropValue().getValue();
-                double newval = VTypeUtil.getValueNumber(vtype).doubleValue();
-
-                if ( !Double.isNaN(newval) ) {
-
-                    if ( newval < min ) {
-                        newval = min;
-                    } else if ( newval > max ) {
-                        newval = max;
-                    }
-
-                    jfx_node.setValue(newval);
-
-                } else {
-//  TODO: CR: do something!!!
-                }
-
-            } finally {
-                updatingValue.set(false);
-            }
-
-        }
-
-    }
-
-    private boolean areZonesVisible ( ) {
-        return model_widget.propShowLoLo().getValue()
-            || model_widget.propShowLow().getValue()
-            || model_widget.propShowHigh().getValue()
-            || model_widget.propShowHiHi().getValue();
     }
 
     @Override
     protected Gauge createJFXNode ( ) throws Exception {
 
-        updateLimits();
-
         knobPosition = model_widget.propKnobPosition().getValue();
         skin = model_widget.propSkin().getValue();
 
-        final Gauge.SkinType skinType;
+        Gauge.SkinType skinType;
 
         switch ( skin ) {
             case THREE_QUARTERS:
@@ -239,31 +129,11 @@ public class MeterRepresentation extends RegionBaseRepresentation<Gauge, MeterWi
                 break;
         }
 
-        Gauge gauge = GaugeBuilder.create()
-                                  .skinType(skinType)
-                                  .prefHeight(model_widget.propHeight().getValue())
-                                  .prefWidth(model_widget.propWidth().getValue())
-                                  //--------------------------------------------------------
-                                  //  Previous properties must be set first.
-                                  //--------------------------------------------------------
-                                  .animated(model_widget.propAnimated().getValue())
-                                  .animationDuration(model_widget.propAnimationDuration().getValue())
-                                  .autoScale(true)
-                                  .checkAreasForValue(false)
-                                  .checkSectionsForValue(false)
-                                  .checkThreshold(false)
-                                  .highlightSections(zonesHighlight)
-                                  .knobPosition(Pos.valueOf(knobPosition.name()))
-                                  .ledVisible(false)
-                                  .maxValue(max)
-                                  .minValue(min)
-                                  .sections(createZones())
-                                  .sectionsVisible(areZonesVisible())
-                                  .tickLabelLocation(TickLabelLocation.INSIDE)
-                                  .title(model_widget.propTitle().getValue())
-                                  .titleColor(JFXUtil.convert(model_widget.propTitleColor().getValue()))
-                                  .value(( max + min ) / 2.0)
-                                  .build();
+        Gauge gauge = super.createJFXNode(skinType);
+
+        gauge.setHighlightSections(zonesHighlight);
+        gauge.setKnobPosition(Pos.valueOf(knobPosition.name()));
+        gauge.setTickLabelLocation(TickLabelLocation.INSIDE);
 
         switch ( skin ) {
             case THREE_QUARTERS:
@@ -280,11 +150,6 @@ public class MeterRepresentation extends RegionBaseRepresentation<Gauge, MeterWi
                 break;
         }
 
-        gauge.animatedProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propAnimated().getValue()) ) {
-                model_widget.propAnimated().setValue(n);
-            }
-        });
         gauge.highlightSectionsProperty().addListener( ( s, o, n ) -> {
             if ( !Objects.equals(n, model_widget.propHighlightZones().getValue()) ) {
                 model_widget.propHighlightZones().setValue(n);
@@ -295,114 +160,64 @@ public class MeterRepresentation extends RegionBaseRepresentation<Gauge, MeterWi
                 model_widget.propKnobPosition().setValue(KnobPosition.valueOf(n.name()));
             }
         });
-        gauge.layoutXProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propX().getValue()) ) {
-                model_widget.propX().setValue(n.intValue());
-            }
-        });
-        gauge.layoutYProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propY().getValue()) ) {
-                model_widget.propY().setValue(n.intValue());
-            }
-        });
-        gauge.prefHeightProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propHeight().getValue()) ) {
-                model_widget.propHeight().setValue(n.intValue());
-            }
-        });
-        gauge.prefWidthProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propWidth().getValue()) ) {
-                model_widget.propWidth().setValue(n.intValue());
-            }
-        });
-        gauge.titleProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propTitle().getValue()) ) {
-                model_widget.propTitle().setValue(n);
-            }
-        });
-        gauge.titleColorProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, JFXUtil.convert(model_widget.propTitleColor().getValue())) ) {
-                model_widget.propTitleColor().setValue(JFXUtil.convert(n));
-            }
-        });
 
         return gauge;
+
+    }
+
+    /**
+     * Creates a new zone with the given parameters.
+     *
+     * @param start The zone's starting value.
+     * @param end   The zone's ending value.
+     * @param name  The zone's name.
+     * @param color The zone's color.
+     * @return A {@link Section} representing the created zone.
+     */
+    @Override
+    protected Section createZone ( double start, double end, String name, Color color ) {
+
+        if ( zonesHighlight ) {
+
+            Section s = new Section(start, end, color.deriveColor(0.0, 1.0, 1.0, 0.2), color);
+
+            s.setText(name);
+
+            return s;
+
+        } else {
+            return super.createZone(start, end, name, color);
+        }
 
     }
 
     @Override
     protected void registerListeners ( ) {
 
-        model_widget.propAnimated().addUntypedPropertyListener(this::behaviorChanged);
-        model_widget.propAnimationDuration().addUntypedPropertyListener(this::behaviorChanged);
-
-        model_widget.propVisible().addUntypedPropertyListener(this::geometryChanged);
-        model_widget.propX().addUntypedPropertyListener(this::geometryChanged);
-        model_widget.propY().addUntypedPropertyListener(this::geometryChanged);
-        model_widget.propWidth().addUntypedPropertyListener(this::geometryChanged);
-        model_widget.propHeight().addUntypedPropertyListener(this::geometryChanged);
+        super.registerListeners();
 
         model_widget.propKnobPosition().addUntypedPropertyListener(this::lookChanged);
         model_widget.propSkin().addUntypedPropertyListener(this::lookChanged);
-        model_widget.propTitle().addUntypedPropertyListener(this::lookChanged);
-        model_widget.propTitleColor().addUntypedPropertyListener(this::lookChanged);
 
         model_widget.propHighlightZones().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propLevelHiHi().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propLevelHight().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propLevelLoLo().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propLevelLow().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propLimitsFromPV().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propShowHiHi().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propShowHigh().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propShowLoLo().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propShowLow().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propMaximum().addUntypedPropertyListener(this::limitsChanged);
-        model_widget.propMinimum().addUntypedPropertyListener(this::limitsChanged);
-
-        if ( toolkit.isEditMode() ) {
-            dirtyValue.checkAndClear();
-        } else {
-            model_widget.runtimePropValue().addPropertyListener(this::valueChanged);
-        }
 
     }
 
-    private Section[] createZones ( ) {
+    @Override
+    protected boolean updateLimits ( ) {
 
-        boolean highlight = model_widget.propHighlightZones().getValue();
-        boolean loloNaN = Double.isNaN(lolo);
-        boolean hihiNaN = Double.isNaN(hihi);
-        List<Section> sections = new ArrayList<>(4);
+        boolean somethingChanged = super.updateLimits();
 
-        if ( !loloNaN ) {
-            sections.add(highlight ? new Section(min, lolo, MAJOR_COLOR_LIGHT, MAJOR_COLOR) : new Section(min, lolo, MAJOR_COLOR));
+        //  Model's values.
+        boolean new_highlight = model_widget.propHighlightZones().getValue();
+
+        if ( zonesHighlight != new_highlight ) {
+            zonesHighlight = new_highlight;
+            somethingChanged = true;
         }
 
-        if ( !Double.isNaN(low) ) {
-            sections.add(highlight ? new Section(loloNaN ? min : lolo, low, MINOR_COLOR_LIGHT, MINOR_COLOR) : new Section(loloNaN ? min : lolo, low, MINOR_COLOR));
-        }
+        return somethingChanged;
 
-        if ( !Double.isNaN(high) ) {
-            sections.add(highlight ? new Section(high, hihiNaN ? max : hihi, MINOR_COLOR_LIGHT, MINOR_COLOR) : new Section(high, hihiNaN ? max : hihi, MINOR_COLOR));
-        }
-
-        if ( !hihiNaN ) {
-            sections.add(highlight ? new Section(hihi, max, MAJOR_COLOR_LIGHT, MAJOR_COLOR) : new Section(hihi, max, MAJOR_COLOR));
-        }
-
-        return sections.toArray(new Section[sections.size()]);
-
-    }
-
-    private void behaviorChanged ( final WidgetProperty<?> property, final Object old_value, final Object new_value ) {
-        dirtyBehavior.mark();
-        toolkit.scheduleUpdate(this);
-    }
-
-    private void geometryChanged ( final WidgetProperty<?> property, final Object old_value, final Object new_value ) {
-        dirtyGeometry.mark();
-        toolkit.scheduleUpdate(this);
     }
 
     private void limitsChanged ( final WidgetProperty<?> property, final Object old_value, final Object new_value ) {
@@ -415,98 +230,6 @@ public class MeterRepresentation extends RegionBaseRepresentation<Gauge, MeterWi
     private void lookChanged ( final WidgetProperty<?> property, final Object old_value, final Object new_value ) {
         dirtyLook.mark();
         toolkit.scheduleUpdate(this);
-    }
-
-    private boolean updateLimits ( ) {
-
-        boolean somethingChanged = false;
-
-        //  Model's values.
-        double new_min = model_widget.propMinimum().getValue();
-        double new_max = model_widget.propMaximum().getValue();
-        double new_lolo = model_widget.propLevelLoLo().getValue();
-        double new_low = model_widget.propLevelLow().getValue();
-        double new_high = model_widget.propLevelHight().getValue();
-        double new_hihi = model_widget.propLevelHiHi().getValue();
-        boolean new_highlight = model_widget.propHighlightZones().getValue();
-
-        if ( model_widget.propLimitsFromPV().getValue() ) {
-
-            //  Try to get display range from PV.
-            final Display display_info = ValueUtil.displayOf(model_widget.runtimePropValue().getValue());
-
-            if ( display_info != null ) {
-                new_min = display_info.getLowerCtrlLimit();
-                new_max = display_info.getUpperCtrlLimit();
-                new_lolo = display_info.getLowerAlarmLimit();
-                new_low = display_info.getLowerWarningLimit();
-                new_high = display_info.getUpperWarningLimit();
-                new_hihi = display_info.getUpperAlarmLimit();
-            }
-
-        }
-
-        if ( !model_widget.propShowLoLo().getValue() ) {
-            new_lolo = Double.NaN;
-        }
-        if ( !model_widget.propShowLow().getValue() ) {
-            new_low = Double.NaN;
-        }
-        if ( !model_widget.propShowHigh().getValue() ) {
-            new_high = Double.NaN;
-        }
-        if ( !model_widget.propShowHiHi().getValue() ) {
-            new_hihi = Double.NaN;
-        }
-
-        //  If invalid limits, fall back to 0..100 range.
-        if ( !( Double.isNaN(new_min) || Double.isNaN(new_max) || new_min < new_max ) ) {
-            new_min = 0.0;
-            new_max = 100.0;
-        }
-
-        if ( Double.compare(min, new_min) != 0 ) {
-            min = new_min;
-            somethingChanged = true;
-        }
-        if ( Double.compare(max, new_max) != 0 ) {
-            max = new_max;
-            somethingChanged = true;
-        }
-        if ( Double.compare(lolo, new_lolo) != 0 ) {
-            lolo = new_lolo;
-            somethingChanged = true;
-        }
-        if ( Double.compare(low, new_low) != 0 ) {
-            low = new_low;
-            somethingChanged = true;
-        }
-        if ( Double.compare(high, new_high) != 0 ) {
-            high = new_high;
-            somethingChanged = true;
-        }
-        if ( Double.compare(hihi, new_hihi) != 0 ) {
-            hihi = new_hihi;
-            somethingChanged = true;
-        }
-        if ( zonesHighlight != new_highlight ) {
-            zonesHighlight = new_highlight;
-            somethingChanged = true;
-        }
-
-        return somethingChanged;
-
-    }
-
-    private void valueChanged ( final WidgetProperty<? extends VType> property, final VType old_value, final VType new_value ) {
-
-        if ( model_widget.propLimitsFromPV().getValue() ) {
-            limitsChanged(null, null, null);
-        }
-
-        dirtyValue.mark();
-        toolkit.scheduleUpdate(this);
-
     }
 
 }
